Replace testGetCredentials with assertion in initCredentials. Remove everything no longer needed from ComputeEngineCloudIT and rename to ComputeClientIT with 1 remaining test.
