+++ pom.xml
@@ -205,7 +205,6 @@
           </systemPropertyVariables>
           <excludes>
             <exclude>**/ComputeEngineCloudWindowsIT.java</exclude>
-            <exclude>**/ComputeEngineCloudIT.java</exclude>
             <exclude>**/ITUtil.java</exclude>
           </excludes>
           <argLine>-Djenkins.test.timeout=${jenkinsRuleTimeout}</argLine>
